+++ docs/source/conf.py
@@ -26,7 +26,7 @@ import scine_autocas
 
 project = 'scine_autocas'
 copyright = """ This code is licensed under the 3-clause BSD license.
-Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
+Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.
 See LICENSE.txt for details.
 """
 author = 'Reiher Group'