+++ scine_autocas/autocas_utils/cas_combination.py
@@ -4,7 +4,7 @@ These functions allow the combination of multiple active spaces along a reaction
 """
 # -*- coding: utf-8 -*-
 __copyright__ = """ This code is licensed under the 3-clause BSD license.
-Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
+Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.
 See LICENSE.txt for details. """
 
 from typing import List, Tuple