+++ scine_autocas/interfaces/molcas/molcas_hdf5_utils.py
@@ -6,7 +6,7 @@ required for active space calculations, without using MOLCASs 'alter' keyword.
 """
 # -*- coding: utf-8 -*-
 __copyright__ = """ This code is licensed under the 3-clause BSD license.
-Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
+Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.
 See LICENSE.txt for details. """
 
 from typing import List, Union