+++ scine_autocas/interfaces/qcmaquis/__init__.py
@@ -5,7 +5,7 @@ mutual information from the qcmaquis hdf5 output.
 """
 # -*- coding: utf-8 -*-
 __copyright__ = """This code is licensed under the 3-clause BSD license.
-Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
+Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.
 See LICENSE.txt for details. """
 
 import numpy as np