+++ scine_autocas/interfaces/qcmaquis/qcmaquis_hdf5_dataset.py
@@ -4,7 +4,7 @@ This module stores all required measurements from a qcmaquis calculation.
 """
 # -*- coding: utf-8 -*-
 __copyright__ = """ This code is licensed under the 3-clause BSD license.
-Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
+Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.
 See LICENSE.txt for details.
 """