+++ scine_autocas/interfaces/qcmaquis/qcmaquis_orbital_rdm_builder.py
@@ -4,7 +4,7 @@ This module implements a class to build orbital RDM from qcmaquis output.
 """
 # -*- coding: utf-8 -*-
 __copyright__ = """This code is licensed under the 3-clause BSD license.
-Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
+Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.
 See LICENSE.txt for details. """
 
 import numpy as np