+++ scine_autocas/tests/test_autocas.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # pylint: disable=C0114, C0115, C0116
 __copyright__ = """ This code is licensed under the 3-clause BSD license.
-Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
+Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.
 See LICENSE.txt for details.
 """
 
@@ -130,6 +130,8 @@ class TestAutocasClasses(unittest.TestCase):
         autocas = Autocas(molecule)
         # to allow predictions in the test. Otherwise it is None, so that it uses data from /dev/urandom or clock
         autocas.large_spaces.seed = 1
+        # TODO: make test pass for max entanglement
+        autocas.large_spaces.average_entanglement = True
         autocas.large_spaces.max_orbitals = 8
         _ = autocas.make_initial_active_space()
         large_cas_occupations, large_cas_indices = autocas.get_large_active_spaces()
@@ -282,6 +284,7 @@ class TestAutocasClasses(unittest.TestCase):
         mutual_information_list = s2_list
         # fmt: on
 
+        autocas.large_spaces.average_entanglement = True
         occupation, s1_entropy, s2_entropy, _ = autocas.collect_entropies(
             index_list, occupation_list, s1_list, s2_list, mutual_information_list
         )