+++ scine_autocas/tests/test_cas_combination.py
@@ -1,6 +1,6 @@
 # pylint: disable=C0114, C0115, C0116
 __copyright__ = """ This code is licensed under the 3-clause BSD license.
-Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
+Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.
 See LICENSE.txt for details.
 """