+++ scine_autocas/tests/test_plots.py
@@ -1,7 +1,7 @@
 # pylint: disable=C0114, C0115, C0116
 # -*- coding: utf-8 -*-
 __copyright__ = """ This code is licensed under the 3-clause BSD license.
-Copyright ETH Zurich, Laboratory of Physical Chemistry, Reiher Group.
+Copyright ETH Zurich, Department of Chemistry and Applied Biosciences, Reiher Group.
 See LICENSE.txt for details. """
 
 import unittest